rest-apis-fastapi/requirements_psycopg.txt
bcrypt==4.0.0
passlib[argon2]>=1.7.4
sqlalchemy>=2.0.45
fastapi>=0.128.0
python-jose[cryptography]>=3.5.0
uvicorn[standard]>=0.40.0
python-multipart>=0.0.21
python-dotenv>=1.2.1

# Additional dependency for psycopg implementation
psycopg2-binary>=2.9.0

